Move SignedAggregateAndProof into own file
